refactor: move file directories
