+++ src/test/java/bowling/domain/player/PlayerTest.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.player;
 
 import static org.assertj.core.api.Assertions.*;
