+++ src/test/java/bowling/domain/result/ResultTest.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.result;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -8,6 +8,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+
+import bowling.domain.score.Score;
+import bowling.domain.score.Scores;
 
 public class ResultTest {
